requirements.txt
amqp==5.3.1
arrow==1.3.0
asgiref==3.8.1
billiard==4.2.1
black==25.1.0
blessed==1.20.0
certifi==2025.1.31
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.6
django-cors-headers==4.7.0
django-environ==0.12.0
django-picklefield==3.2
django-redis==5.4.0
django-rq==3.0.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.7
gunicorn==23.0.0
huey==2.5.2
inflection==0.5.1
kombu==5.4.2
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pip-system-certs==4.0
pipdeptree==2.25.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
rq==2.1.0
six==1.17.0
sqlparse==0.5.3
types-python-dateutil==2.9.0.20241206
tzdata==2025.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
wrapt==1.17.2
